requirements.txt
fastapi 
uvicorn
requests
ollama
sqlalchemy
pydantic
python-dotenv
psycopg2
asyncpg
greenlet